Moved clickjacking decorator tests into decorators/test_clickjacking.py.
This also adds extra assertions.
